pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-authx"
version = "0.1.0"
description = "Reusable Django authentication package with email-based authentication, OTP verification, password reset, and JWT support."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Alfer" }
]
keywords = ["django", "authentication", "otp", "jwt", "email-auth", "password-reset"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "Django>=4.2",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=5.3",
    "python-decouple>=3.8",
]

[project.urls]
Homepage = "https://example.com"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]